pyproject.toml
[tool.black]
line-length = 88
target-version = ["py312"]
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]  # flake8 오류 + import 정리
ignore = ["E501"]
exclude = ["alembic"]

[tool.ruff.format]
quote-style = "double"

[project]
name = "Runna-backend"
version = "0.1.0"
description = "Run your code with serverless"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "black>=25.11.0",
    "esprima>=4.0.1",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "redis>=7.1.0",
    "ruff>=0.14.7",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.12.0",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "kubernetes>=34.1.0",
]
